Revert meteor-promise addition, for now.
Webpack-ifying this package when it depends on meteor-promise pulls in a
lot more code than we need, and it's the server implementation of
meteor-promise, too. Need to think harder about how to do this. Maybe
meteor-promise can expect Promise support to be provided from another
source, like this package, rather than providing its own polyfill?